rename: 도메인 클래스 패키징 변경
