+++ src/main/java/domain/ladder/LadderFactory.java
@@ -1,7 +1,4 @@
-package factory;
-
-import domain.Ladder;
-import domain.Line;
+package domain.ladder;
 
 import java.util.ArrayList;
 import java.util.List;
