+++ src/main/java/domain/ladder/Line.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import java.util.ArrayList;
 import java.util.Collections;
