+++ src/main/java/domain/players/PlayerName.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.players;
 
 public class PlayerName {
