+++ src/test/java/domain/LineTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.ladder.Line;
+import domain.ladder.Point;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
